other major diff is moving files into src/lib
