+++ src/index.ts
@@ -1,15 +1,15 @@
 export * from './codecs';
 export * from './enums';
-export * from './readTranscendYaml';
-export * from './graphql/syncConfigurationToTranscend';
-export * from './mergeTranscendInputs';
-export * from './requests';
+export * from './lib/readTranscendYaml';
+export * from './lib/graphql/syncConfigurationToTranscend';
+export * from './lib/mergeTranscendInputs';
+export * from './lib/requests';
 export * from './constants';
-export * from './consent-manager';
-export * from './graphql';
-export * from './helpers';
-export * from './cron';
-export * from './api-keys';
-export * from './ai';
-export * from './data-inventory';
-export * from './manual-enrichment';
+export * from './lib/consent-manager';
+export * from './lib/graphql';
+export * from './lib/helpers';
+export * from './lib/cron';
+export * from './lib/api-keys';
+export * from './lib/ai';
+export * from './lib/data-inventory';
+export * from './lib/manual-enrichment';
