+++ src/lib/cli-request-enricher-restart.ts
@@ -3,13 +3,13 @@
 import yargs from 'yargs-parser';
 import colors from 'colors';
 
-import { logger } from './logger';
+import { logger } from '../logger';
 import { splitCsvToList, bulkRetryEnrichers } from './requests';
 import {
   RequestAction,
   RequestEnricherStatus,
 } from '@transcend-io/privacy-types';
-import { DEFAULT_TRANSCEND_API } from './constants';
+import { DEFAULT_TRANSCEND_API } from '../constants';
 
 /**
  * Run a command to bulk restart a specific enricher
