+++ src/lib/cli-request-export.ts
@@ -3,11 +3,11 @@
 import yargs from 'yargs-parser';
 import colors from 'colors';
 
-import { logger } from './logger';
+import { logger } from '../logger';
 import uniq from 'lodash/uniq';
 import { RequestAction, RequestStatus } from '@transcend-io/privacy-types';
 import { splitCsvToList, pullPrivacyRequests } from './requests';
-import { DEFAULT_TRANSCEND_API } from './constants';
+import { DEFAULT_TRANSCEND_API } from '../constants';
 import { writeCsv } from './cron';
 
 /**
