+++ src/lib/cli-request-notify-additional-time.ts
@@ -3,13 +3,13 @@
 import yargs from 'yargs-parser';
 import colors from 'colors';
 
-import { logger } from './logger';
+import { logger } from '../logger';
 import { RequestAction } from '@transcend-io/privacy-types';
 import {
   splitCsvToList,
   notifyPrivacyRequestsAdditionalTime,
 } from './requests';
-import { DEFAULT_TRANSCEND_API } from './constants';
+import { DEFAULT_TRANSCEND_API } from '../constants';
 
 /**
  * Notify requests that more time is needed
