+++ src/lib/cli-request-upload.ts
@@ -3,9 +3,9 @@
 import yargs from 'yargs-parser';
 import colors from 'colors';
 
-import { logger } from './logger';
+import { logger } from '../logger';
 import { splitCsvToList, uploadPrivacyRequestsFromCsv } from './requests';
-import { DEFAULT_TRANSCEND_API } from './constants';
+import { DEFAULT_TRANSCEND_API } from '../constants';
 
 /**
  * Upload a CSV of Privacy Requests.
