+++ src/lib/cli-skip-preflight-jobs.ts
@@ -3,9 +3,9 @@
 import yargs from 'yargs-parser';
 import colors from 'colors';
 
-import { logger } from './logger';
+import { logger } from '../logger';
 import { skipPreflightJobs, splitCsvToList } from './requests';
-import { DEFAULT_TRANSCEND_API } from './constants';
+import { DEFAULT_TRANSCEND_API } from '../constants';
 
 /**
  * Given an enricher ID, skip all active enrichment jobs that are open for that enricher
