+++ src/lib/cli-skip-request-data-silos.ts
@@ -3,9 +3,9 @@
 import yargs from 'yargs-parser';
 import colors from 'colors';
 
-import { logger } from './logger';
+import { logger } from '../logger';
 import { skipRequestDataSilos, splitCsvToList } from './requests';
-import { DEFAULT_TRANSCEND_API } from './constants';
+import { DEFAULT_TRANSCEND_API } from '../constants';
 import { RequestStatus } from '@transcend-io/privacy-types';
 
 /**
