+++ src/lib/consent-manager/uploadCookiesFromCsv.ts
@@ -1,11 +1,11 @@
 import colors from 'colors';
-import { logger } from '../logger';
+import { logger } from '../../logger';
 import { ConsentTrackerStatus } from '@transcend-io/privacy-types';
 import { buildTranscendGraphQLClient, syncCookies } from '../graphql';
 import { readCsv } from '../requests/readCsv';
-import { CookieInput, CookieCsvInput } from '../codecs';
+import { CookieInput, CookieCsvInput } from '../../codecs';
 import { splitCsvToList } from '../requests';
-import { DEFAULT_TRANSCEND_API } from '../constants';
+import { DEFAULT_TRANSCEND_API } from '../../constants';
 
 const OMIT_COLUMNS = [
   'ID',
